Move \Codeception\Lib\Interfaces\CustomCommand to Codeception\CustomCommandInterface
