+++ tests/data/register_command/examples/MyCustomCommand.php
@@ -8,12 +8,12 @@
 namespace Codeception\Command;
 
 use \Symfony\Component\Console\Command\Command;
-use \Codeception\Lib\Interfaces\CustomCommand;
+use \Codeception\CustomCommandInterface;
 use \Symfony\Component\Console\Input\InputOption;
 use \Symfony\Component\Console\Input\InputInterface;
 use \Symfony\Component\Console\Output\OutputInterface;
 
-class MyCustomCommand extends Command implements CustomCommand
+class MyCustomCommand extends Command implements CustomCommandInterface
 {
 
     use Shared\FileSystem;
