+++ tests/support/_generated/CoverGuyActions.php
@@ -1,4 +1,4 @@
-<?php  //[STAMP] ccc2227a1b79d46c35e7231e4f8015d3
+<?php  //[STAMP] 008cc27ead2e55b4e5d0fe0e0dd174ee
 namespace _generated;
 
 // This class was automatically generated by build task
